requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Faker==37.8.0
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.1
python-slugify==8.0.4
python-dotenv==1.1.1
requests==2.32.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
psycopg2-binary